requirements.txt
flask==2.3.3
transformers==4.35.2
torch==2.1.0
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.0.1
speechrecognition==3.10.0
pyaudio==0.2.13
gTTS==2.5.0